Move client daemon log to a common place
In multi-user environment, client daemon will be lauched by first user,
the log located on the home of that first user is not suitable.
Meanwhile, on the next pc reboot, the so called `first user` possible be
another user, which makes us not easy to find the correct log location.

Move the daemon log to /tmp/fc will makes things easy to check.

Signed-off-by: Larry Shen <larry.shen@nxp.com>