server_daemon: fix daemon clean failure when in container
Usually, the prctl will handle the signal sent to child if parent process exit.
But in container, the parent process act as PID1, once it exits, the os did not
send any signal to child process, this makes etcd not be aware of server leave
until the lock expire.

`docker stop` will send SIGTERM to PID1, so let parent monitor signal 15 to handle
container stop, to let etcd be aware of server leave in time in docker deployment.

Signed-off-by: Larry Shen <larry.shen@nxp.com>